requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
psycopg2-binary
psycopg2
bcrypt
fastapi-jwt-auth
tortoise-orm
# jwt
PyJWT
python-jose
passlib
python-multipart
python-dotenv
pymongo
dist/utils-1.0.tar.gz